java: Set release version to 17 (#64)
* Set default release version to 17
* Enable Error prone with Nullaway
* Enable Checkstyle

+++ java/src/main/java/io/cucumber/prettyformatter/PrettyReportData.java
@@ -215,7 +215,7 @@ final class PrettyReportData {
         return query.findPickleBy(testCaseStarted);
     }
 
-    Optional<Long> findLineOf(Pickle pickle) {
+    Optional<Integer> findLineOf(Pickle pickle) {
         return query.findLocationOf(pickle).map(Location::getLine);
     }
